+++ src/Contracts/TrainableContract.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace SavvyAI\Contracts;
+
+use Illuminate\Contracts\Database\Eloquent\Builder;
+
+interface TrainableContract
+{
+    public function getStatementRepository(): Builder;
+}